// frontend/src/codegen/model/sortBy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flathub API
 * OpenAPI spec version: 0.1.0
 */

export type SortBy = (typeof SortBy)[keyof typeof SortBy]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SortBy = {
  trending: "trending",
  installs_last_month: "installs_last_month",
} as const